Add files via upload
	EDA (Exploratory Data Analysis of IPL Matches)
      1 January 2023 to 15 March 2023

      Library in python: -Panda
     Objectives of this project: 
I am used IPL 2008-2017 matches dataset in this project. 
To find the team that won the most number of matches in a season.
To find the team that lost the most number of matches in a season.
Does winning toss increases the chances of victory.
To find the player with the most player of the match awards.
To find the city that hosted the maximum number of IPL matches.
To find the most winning team for each season.
To find the on-field umpire with the maximum number of IPL matches.
To find the biggest victories in IPL while defending a total and while chasing a total.