Add files via upload
	Tic Tac Toe Game  
1 October 2022 to 30 December 2022.

It is a two-player game. In this game, there is a board with 3×3 squares. In this game    a player can choose between two symbols with his opponent, the usual games use “X” and “O”. If the first player chooses “X”, then the second player must play with “O” and vice versa.

Programming language: - Python
Library in python: - tkinter